+++ tests/test_updated_esdl_post_process.py
@@ -258,10 +258,14 @@ class TestUpdatedESDL(TestCase):
 
                 # Check pipe diameter
                 if len(fnmatch.filter([energy_system.instance[0].area.asset[ii].id], "Pipe*")) == 1:
-                    if asset_name in ["Pipe1", "Pipe1_cold_pipe", "Pipe4", "Pipe4_ret"]:
+                    if asset_name in ["Pipe1", "Pipe1_ret"]:
                         np.testing.assert_array_equal(
-                            energy_system.instance[0].area.asset[ii].diameter.name, "DN150"
+                            energy_system.instance[0].area.asset[ii].diameter.name, "DN250"
                         )  # original pipe DN400 being sized
+                    elif asset_name in ["Pipe4", "Pipe4_ret"]:
+                        np.testing.assert_array_equal(
+                            energy_system.instance[0].area.asset[ii].diameter.name, "DN200"
+                        )  # original pipe DN900 being sized
                     elif asset_name not in ["Pipe5", "Pipe5_ret"]:
                         np.testing.assert_array_equal(
                             energy_system.instance[0].area.asset[ii].diameter.name, "DN400"